Add the ability to do an exact search
Why?

Often we know a unique result will be returned given a certain set of criteria but we need to do a full search and then get the first element of the returned array. There are also several custom implementations of these types of searches so creating a generic function.