Initial support for multiple concurrent SQLAlchemy engines
SQLAlchemy's `sessionmaker` supports a custom session type that, in turn, can override
`get_bind()` to control which engine it uses. Since the SQLAlchemy engine encompasses
the underlying Postgres connection and credentials, this approach allows us to use
different connections based on runtime context.

We can combine this notion with the `microcosm` "scoped factory" feature to allow multiple
configurations of the same entity to be managed simultaneously, with shared defines.

There are two caveats:
 1. The strategy used here to implement `get_bind()` isn't generic enough to meet our
    current anticipated use cases (e.g. connection-switching based on user auth)
 2. The scoped factory feature uses object proxies to manage object graph references;
    we have to be careful to unwrap these references to avoid breaking SQLAlchemy internals
    that performance object identity comparisons.