Support temporary tables, especially for bulk upsert
One of the most effective ways to load large amounts of postgres data is to
use a bulk upsert. However, in the likely event that you want to know which
of the original records were inserted (vs ignored on conflict), it's helpful
to have a temporary table to insert from and then query for id matches.

Usage:

    # create a transient table that copies Target (using "on commit drop")
    with transient(Target) as transient_table:

        # insert many Target instances
        transient_table.insert_many([Target(...), Target(...), ...])

        # copy records from transient table into Target
        transient_table.copy_into(Target)

        # return the Target instances that were inserted
        return transient_table.select_from(TargetModel)