Remove pagination out of filter (#51)
Why?

The count call to the store should not have pagination factored in. This is typically done at the CRUDStoreAdaptor level in microcosm-flask (by just avoiding passing in limit/offset to the store's count function) but more accurately should be done in the store. Also makes it consistent with microcosm-sqlite