Expose the synchronize session delete param.
Why? SqlAlchemy `delete` attemps to synchronize the session before the
commit. It does so by taking a middle ground between performance and
convenience, but it also exposes other strategies based on the use case.
This PR exposes those other strategies for other use cases that may
arise.

The default `evaluate` is nice, but the way the store is written
requires overwriting the built-in `_delete` method to gain access to the
other `sqlalchemy` session synch strategies. This matters because the
other strategies (no strategy and `fetch`) are more efficient and more
convenient respectively.