Add option to cascade table truncation in `operations#recreate_all()`.
Occasionally we will have circumstances where there are cyclic foreign
key constraints on tables.

For instance, a 1 to many relationship between UserIdentity and User.
UserIdentity has a `created_by`, which is a f.k. on `user.id`
User has a `user_identity_id`, which is a f.k. on `user_identity.id`

SqlAlchemy recommends naming Foreign Keys to get around cyclic dependencies,
but the resolution of dependent tables is still not perfect. It so happens
that sometimes the order MetaData.sorted_tables returns tables in is partially
alphabetical. In this instance, calling `.recreate_all()` in test suites
will yield a ForeignKeyViolation, because there are still references to the
table you are trying to truncate in other tables.

In our example above, the reversed sorted tables (i.e. order of deletion)
indicated UserIdentity was first, and then User. You cannot delete a
UserIdentity whilst a User still references it in the `user_identity_id` field,
thus the operation fails.

In this instance, we want to force a cascading delete. We can achieve this
by passing in the flag in this commit to `recreate_all()`. This is a faster
alternative to dropping and recreating all the tables from scratch in between
test runs.