Explicitly enable `restore_load_context` for encrypted entities
This directive informs SQLA that it should restore the original context
(i.e. runid) for a for a given entity, instead of treating it as a
different context.

This prevents a warning from being erroneously emitted in the following
scenario:
* Unencrypted Parent class
* Encrypted polymorphic subclass of Parent

When the parent class is loaded, SQLA doesn't appear to take into
account any relationships of the relevant subclasses, meaning that it
needs to emit extra selects to fetch those attached entities once the
subclass has been fully loaded.