Prepare library for aws-encryption-sdk 2.0
Replaces deprecated usage of encrypt/decrypt with the new dedicated
EncryptionSDKClient.

The main driver change here appears to have been with "key commitment" (see ref).
The recommended change here is to use the client explicitly configured to be
backwards compatible

This should hopefully prepare users of the library to safely upgrade to
2.0, once these changes have been verified as working correctly.

Ref:
https://docs.aws.amazon.com/encryption-sdk/latest/developer-guide/migration-guide.html
https://docs.aws.amazon.com/encryption-sdk/latest/developer-guide/migration.html
https://docs.aws.amazon.com/encryption-sdk/latest/developer-guide/concepts.html#key-commitment