Replace usage of KMSMasterKeyProvider
This class was explicitly deprecated and removed in 2.x, to be replaced
with two different kinds of KeyProviders:
* StrictAwsKmsMasterKeyProvider
* DiscoveryAwsKmsMasterKeyProvider

The strict provider will continue to work as the previous provider had
done, encrypting new rows with the explicitly-provided keys.

Decryption, however, will use the discovery provider to decrypt values
in order to keep working with older accounts and values not explicitly
listed. This may change in the future, as it may be preferable to use a
strict provider on both ends. This will require, however, all rows to be
vetted to ensure all keys are accounted for.

https://docs.aws.amazon.com/encryption-sdk/latest/developer-guide/migrate-mkps-v2.html